mko rest

+++ tnave copy/get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: tnave <tnave@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/26 16:01:54 by tnave             #+#    #+#             */
-/*   Updated: 2021/01/27 15:13:07 by tnave            ###   ########.fr       */
+/*   Updated: 2021/01/27 11:02:21 by tnave            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@
 size_t	ft_strlen(char const *str);
 char	*ft_strjoin(char *s1, char *s2);
 char	*ft_reste(char *buff);
-size_t	ft_get_n(char *temp);
+int		ft_get_n(char *temp);
 char	*get_first_line(int fd, char **line, char *buff, char *temp);
 char 	*get_temp_line(char **line, char *temp);
 int		get_next_line(int fd, char **line);
